initialize self.valve_test to be None.
Changed disable/enable motor controls to be disable/enable_gui and added inputs as well as buttons.

Put a check in start_test_button_handler so that a second valve test cannot be started over another valve test that is running.

Added check at cleanup to see if a valve test is running when the window is closing.